+++ src/GR.Extensions/GR.Ecommerce.Extension/GR.ECommerce.Abstractions/Models/Product.cs
@@ -84,17 +84,17 @@ namespace GR.ECommerce.Abstractions.Models
         /// <summary>
         /// Discount price
         /// </summary>
-        public double DiscountPrice => (PriceWithoutDiscount * ProductDiscount?.Discount?.Percentage ?? 0) / 100;
+        public decimal DiscountPrice => (PriceWithoutDiscount * ProductDiscount?.Discount?.Percentage ?? 0) / 100;
 
         /// <summary>
         /// Current price
         /// </summary>
-        public virtual double PriceWithoutDiscount => ProductPrices?.OrderBy(x => x.Created).LastOrDefault()?.Price ?? 0;
+        public virtual decimal PriceWithoutDiscount => ProductPrices?.OrderBy(x => x.Created).LastOrDefault()?.Price ?? 0;
 
         /// <summary>
         /// Final price with discount
         /// </summary>
-        public virtual double PriceWithDiscount => PriceWithoutDiscount - DiscountPrice;
+        public virtual decimal PriceWithDiscount => PriceWithoutDiscount - DiscountPrice;
 
         /// <summary>
         /// Publish state of product
